More records - moved stuff to constants to get rid of cyclic imports
Was getting AttributeError: 'module' object has no attribute 'init'
Diffs from vanilla rip (excluding the move)
@@ -27,4 +27,2 @@
 import struct
-import bash
-import bash.brec
 from . import esp
@@ -32,4 +30,3 @@
 #--Mod I/O
-from ...bolt import StateError, Flags
-import bolt
+from ...bolt import StateError, Flags, BoltError, sio
 from ...brec import MelRecord, BaseRecordHeader, ModError, MelStructs, null3, \
@@ -38,4 +35,4 @@
     MelFids
-import bush
-from game.oblivion import allConditions, fid1Conditions, fid2Conditions
+from ...bush import genericAVEffects, mgef_school, mgef_basevalue, actorValues
+from oblivion_const import allConditions, fid1Conditions, fid2Conditions

@@ -105,3 +102,3 @@ def mergeFilter(self,modSet):
 #------------------------------------------------------------------------------
-class MelBipedFlags(bash.bolt.Flags):
+class MelBipedFlags(Flags):
     """Biped flags element. Includes biped flag set by default."""
@@ -150,3 +147,3 @@ def loadData(self,record,ins,type,size,readId):
         if ifunc not in allConditions:
-            raise bolt.BoltError(u'Unknown condition function: %d' % ifunc)
+            raise BoltError(u'Unknown condition function: %d' % ifunc)
         form1 = 'iI'[ifunc in fid1Conditions]
@@ -191,3 +188,3 @@ class MelEffects(MelGroups):
     #--Class Data
-    seFlags = bolt.Flags(0x0L,bolt.Flags.getNames('hostile'))
+    seFlags = Flags(0x0L,Flags.getNames('hostile'))
     class MelEffectsScit(MelStruct):
@@ -239,3 +236,3 @@ def getEffects(self):
         effects = []
-        avEffects = bush.genericAVEffects
+        avEffects = genericAVEffects
         effectsAppend = effects.append
@@ -248,3 +245,3 @@ def getEffects(self):

-    def getSpellSchool(self,mgef_school=bush.mgef_school):
+    def getSpellSchool(self,mgef_school=mgef_school):
         """Returns the school based on the highest cost spell effect."""
@@ -253,3 +250,3 @@ def getSpellSchool(self,mgef_school=bush.mgef_school):
             school = mgef_school[effect.name]
-            effectValue = bush.mgef_basevalue[effect.name]
+            effectValue = mgef_basevalue[effect.name]
             if effect.magnitude:
@@ -266,7 +263,7 @@ def getEffectsSummary(self,mgef_school=None,mgef_name=None):
         """Return a text description of magic effects."""
-        mgef_school = mgef_school or bush.mgef_school
-        mgef_name = mgef_name or bush.mgef_name
-        with bolt.sio() as buff:
-            avEffects = bush.genericAVEffects
-            aValues = bush.actorValues
+        mgef_school = mgef_school or mgef_school
+        mgef_name = mgef_name or mgef_name
+        with sio() as buff:
+            avEffects = genericAVEffects
+            aValues = actorValues
             buffWrite = buff.write
@@ -274,3 +271,3 @@ def getEffectsSummary(self,mgef_school=None,mgef_name=None):
                 school = self.getSpellSchool(mgef_school)
-                buffWrite(bush.actorValues[20+school] + u'\n')
+                buffWrite(actorValues[20+school] + u'\n')
             for index,effect in enumerate(self.effects):
@@ -364,2 +361 @@ def mapFids(self,record,function,save=False):
                 if save: scrxen[index] = (isFid,result)
-
