+++ Mopy/bash/game/oblivion/__init__.py
@@ -27,7 +27,6 @@
 
 from oblivion_const import bethDataFiles, allBethFiles
 from ... import brec
-from ... import bush
 from ...brec import *
 from ...bolt import Flags, DataDict, StateError
 
@@ -303,182 +302,6 @@ ignoreDataDirs = set((
     u'OBSE\\Plugins\\ComponentDLLs\\CSE',
     u'LSData'
     ))
-
-# Function Info ---------------------------------------------------------------
-conditionFunctionData = ( #--0: no param; 1: int param; 2: formid param
-    (153, 'CanHaveFlames', 0, 0),
-    (127, 'CanPayCrimeGold', 0, 0),
-    ( 14, 'GetActorValue', 1, 0),
-    ( 61, 'GetAlarmed', 0, 0),
-    (190, 'GetAmountSoldStolen', 0, 0),
-    (  8, 'GetAngle', 1, 0),
-    ( 81, 'GetArmorRating', 0, 0),
-    (274, 'GetArmorRatingUpperBody', 0, 0),
-    ( 63, 'GetAttacked', 0, 0),
-    (264, 'GetBarterGold', 0, 0),
-    (277, 'GetBaseActorValue', 1, 0),
-    (229, 'GetClassDefaultMatch', 0, 0),
-    ( 41, 'GetClothingValue', 0, 0),
-    (122, 'GetCrime', 2, 1),
-    (116, 'GetCrimeGold', 0, 0),
-    (110, 'GetCurrentAIPackage', 0, 0),
-    (143, 'GetCurrentAIProcedure', 0, 0),
-    ( 18, 'GetCurrentTime', 0, 0),
-    (148, 'GetCurrentWeatherPercent', 0, 0),
-    (170, 'GetDayOfWeek', 0, 0),
-    ( 46, 'GetDead', 0, 0),
-    ( 84, 'GetDeadCount', 2, 0),
-    (203, 'GetDestroyed', 0, 0),
-    ( 45, 'GetDetected', 2, 0),
-    (180, 'GetDetectionLevel', 2, 0),
-    ( 35, 'GetDisabled', 0, 0),
-    ( 39, 'GetDisease', 0, 0),
-    ( 76, 'GetDisposition', 2, 0),
-    (  1, 'GetDistance', 2, 0),
-    (215, 'GetDoorDefaultOpen', 0, 0),
-    (182, 'GetEquipped', 2, 0),
-    ( 73, 'GetFactionRank', 2, 0),
-    ( 60, 'GetFactionRankDifference', 2, 2),
-    (128, 'GetFatiguePercentage', 0, 0),
-    (288, 'GetFriendHit', 2, 0),
-    (160, 'GetFurnitureMarkerID', 0, 0),
-    ( 74, 'GetGlobalValue', 2, 0),
-    ( 48, 'GetGold', 0, 0),
-    ( 99, 'GetHeadingAngle', 2, 0),
-    (318, 'GetIdleDoneOnce', 0, 0),
-    (338, 'GetIgnoreFriendlyHits', 0, 0),
-    ( 67, 'GetInCell', 2, 0),
-    (230, 'GetInCellParam', 2, 2),
-    ( 71, 'GetInFaction', 2, 0),
-    ( 32, 'GetInSameCell', 2, 0),
-    (305, 'GetInvestmentGold', 0, 0),
-    (310, 'GetInWorldspace', 2, 0),
-    ( 91, 'GetIsAlerted', 0, 0),
-    ( 68, 'GetIsClass', 2, 0),
-    (228, 'GetIsClassDefault', 2, 0),
-    ( 64, 'GetIsCreature', 0, 0),
-    (161, 'GetIsCurrentPackage', 2, 0),
-    (149, 'GetIsCurrentWeather', 2, 0),
-    (237, 'GetIsGhost', 0, 0),
-    ( 72, 'GetIsID', 2, 0),
-    (254, 'GetIsPlayableRace', 0, 0),
-    (224, 'GetIsPlayerBirthsign', 2, 0),
-    ( 69, 'GetIsRace', 2, 0),
-    (136, 'GetIsReference', 2, 0),
-    ( 70, 'GetIsSex', 1, 0),
-    (246, 'GetIsUsedItem', 2, 0),
-    (247, 'GetIsUsedItemType', 1, 0),
-    ( 47, 'GetItemCount', 2, 0),
-    (107, 'GetKnockedState', 0, 0),
-    ( 80, 'GetLevel', 0, 0),
-    ( 27, 'GetLineOfSight', 2, 0),
-    (  5, 'GetLocked', 0, 0),
-    ( 65, 'GetLockLevel', 0, 0),
-    (320, 'GetNoRumors', 0, 0),
-    (255, 'GetOffersServicesNow', 0, 0),
-    (157, 'GetOpenState', 0, 0),
-    (193, 'GetPCExpelled', 2, 0),
-    (199, 'GetPCFactionAttack', 2, 0),
-    (195, 'GetPCFactionMurder', 2, 0),
-    (197, 'GetPCFactionSteal', 2, 0),
-    (201, 'GetPCFactionSubmitAuthority', 2, 0),
-    (249, 'GetPCFame', 0, 0),
-    (132, 'GetPCInFaction', 2, 0),
-    (251, 'GetPCInfamy', 0, 0),
-    (129, 'GetPCIsClass', 2, 0),
-    (130, 'GetPCIsRace', 2, 0),
-    (131, 'GetPCIsSex', 1, 0),
-    (312, 'GetPCMiscStat', 1, 0),
-    (225, 'GetPersuasionNumber', 0, 0),
-    ( 98, 'GetPlayerControlsDisabled', 0, 0),
-    (365, 'GetPlayerInSEWorld',0,0),
-    (362, 'GetPlayerHasLastRiddenHorse', 0, 0),
-    (  6, 'GetPos', 1, 0),
-    ( 56, 'GetQuestRunning', 2, 0),
-    ( 79, 'GetQuestVariable', 2, 1),
-    ( 77, 'GetRandomPercent', 0, 0),
-    (244, 'GetRestrained', 0, 0),
-    ( 24, 'GetScale', 0, 0),
-    ( 53, 'GetScriptVariable', 2, 1),
-    ( 12, 'GetSecondsPassed', 0, 0),
-    ( 66, 'GetShouldAttack', 2, 0),
-    (159, 'GetSitting', 0, 0),
-    ( 49, 'GetSleeping', 0, 0),
-    ( 58, 'GetStage', 2, 0),
-    ( 59, 'GetStageDone', 2, 1),
-    ( 11, 'GetStartingAngle', 1, 0),
-    ( 10, 'GetStartingPos', 1, 0),
-    ( 50, 'GetTalkedToPC', 0, 0),
-    (172, 'GetTalkedToPCParam', 2, 0),
-    (361, 'GetTimeDead', 0, 0),
-    (315, 'GetTotalPersuasionNumber', 0, 0),
-    (144, 'GetTrespassWarningLevel', 0, 0),
-    (242, 'GetUnconscious', 0, 0),
-    (259, 'GetUsedItemActivate', 0, 0),
-    (258, 'GetUsedItemLevel', 0, 0),
-    ( 40, 'GetVampire', 0, 0),
-    (142, 'GetWalkSpeed', 0, 0),
-    (108, 'GetWeaponAnimType', 0, 0),
-    (109, 'GetWeaponSkillType', 0, 0),
-    (147, 'GetWindSpeed', 0, 0),
-    (154, 'HasFlames', 0, 0),
-    (214, 'HasMagicEffect', 2, 0),
-    (227, 'HasVampireFed', 0, 0),
-    (353, 'IsActor', 0, 0),
-    (314, 'IsActorAVictim', 0, 0),
-    (313, 'IsActorEvil', 0, 0),
-    (306, 'IsActorUsingATorch', 0, 0),
-    (280, 'IsCellOwner', 2, 2),
-    (267, 'IsCloudy', 0, 0),
-    (150, 'IsContinuingPackagePCNear', 0, 0),
-    (163, 'IsCurrentFurnitureObj', 2, 0),
-    (162, 'IsCurrentFurnitureRef', 2, 0),
-    (354, 'IsEssential', 0, 0),
-    (106, 'IsFacingUp', 0, 0),
-    (125, 'IsGuard', 0, 0),
-    (282, 'IsHorseStolen', 0, 0),
-    (112, 'IsIdlePlaying', 0, 0),
-    (289, 'IsInCombat', 0, 0),
-    (332, 'IsInDangerousWater', 0, 0),
-    (300, 'IsInInterior', 0, 0),
-    (146, 'IsInMyOwnedCell', 0, 0),
-    (285, 'IsLeftUp', 0, 0),
-    (278, 'IsOwner', 2, 0),
-    (176, 'IsPCAMurderer', 0, 0),
-    (175, 'IsPCSleeping', 0, 0),
-    (171, 'IsPlayerInJail', 0, 0),
-    (358, 'IsPlayerMovingIntoNewSpace', 0, 0),
-    (339, 'IsPlayersLastRiddenHorse', 0, 0),
-    (266, 'IsPleasant', 0, 0),
-    ( 62, 'IsRaining', 0, 0),
-    (327, 'IsRidingHorse', 0, 0),
-    (287, 'IsRunning', 0, 0),
-    (103, 'IsShieldOut', 0, 0),
-    (286, 'IsSneaking', 0, 0),
-    ( 75, 'IsSnowing', 0, 0),
-    (223, 'IsSpellTarget', 2, 0),
-    (185, 'IsSwimming', 0, 0),
-    (141, 'IsTalking', 0, 0),
-    (265, 'IsTimePassing', 0, 0),
-    (102, 'IsTorchOut', 0, 0),
-    (145, 'IsTrespassing', 0, 0),
-    (329, 'IsTurnArrest', 0, 0),
-    (111, 'IsWaiting', 0, 0),
-    (101, 'IsWeaponOut', 0, 0),
-    (309, 'IsXBox', 0, 0),
-    (104, 'IsYielding', 0, 0),
-    ( 36, 'MenuMode', 1, 0),
-    ( 42, 'SameFaction', 2, 0),
-    (133, 'SameFactionAsPC', 0, 0),
-    ( 43, 'SameRace', 2, 0),
-    (134, 'SameRaceAsPC', 0, 0),
-    ( 44, 'SameSex', 2, 0),
-    (135, 'SameSexAsPC', 0, 0),
-    (323, 'WhichServiceMenu', 0, 0),
-    )
-allConditions = set(entry[0] for entry in conditionFunctionData)
-fid1Conditions = set(entry[0] for entry in conditionFunctionData if entry[2] == 2)
-fid2Conditions = set(entry[0] for entry in conditionFunctionData if entry[3] == 2)
 
 #--List of GMST's in the main plugin (Oblivion.esm) that have 0x00000000
 #  as the form id.  Any GMST as such needs it Editor Id listed here.
@@ -1679,262 +1502,11 @@ class esp:
 
 from records import * # MUST BE HERE otherwise all hell breaks loose
 #------------------------------------------------------------------------------
-class MelConditions(MelStructs):
-    """Represents a set of quest/dialog conditions. Difficulty is that FID state
-    of parameters depends on function index."""
-    def __init__(self):
-        """Initialize."""
-        MelStructs.__init__(self,'CTDA','B3sfIii4s','conditions',
-            'operFlag',('unused1',null3),'compValue','ifunc','param1','param2',('unused2',null4))
-
-    def getLoaders(self,loaders):
-        """Adds self as loader for type."""
-        loaders[self.subType] = self
-        loaders['CTDT'] = self #--Older CTDT type for ai package records.
-
-    def getDefault(self):
-        """Returns a default copy of object."""
-        target = MelStructs.getDefault(self)
-        target.form12 = 'ii'
-        return target
-
-    def hasFids(self,formElements):
-        """Include self if has fids."""
-        formElements.add(self)
-
-    def loadData(self,record,ins,type,size,readId):
-        """Reads data from ins into record attribute."""
-        if type == 'CTDA' and size != 24:
-            raise ModSizeError(ins.inName,readId,24,size,True)
-        if type == 'CTDT' and size != 20:
-            raise ModSizeError(ins.inName,readId,20,size,True)
-        target = MelObject()
-        record.conditions.append(target)
-        target.__slots__ = self.attrs
-        unpacked1 = ins.unpack('B3sfI',12,readId)
-        (target.operFlag,target.unused1,target.compValue,ifunc) = unpacked1
-        #--Get parameters
-        if ifunc not in allConditions:
-            raise bolt.BoltError(u'Unknown condition function: %d' % ifunc)
-        form1 = 'iI'[ifunc in fid1Conditions]
-        form2 = 'iI'[ifunc in fid2Conditions]
-        form12 = form1+form2
-        unpacked2 = ins.unpack(form12,8,readId)
-        (target.param1,target.param2) = unpacked2
-        if size == 24:
-            target.unused2 = ins.read(4)
-        else:
-            target.unused2 = null4
-        (target.ifunc,target.form12) = (ifunc,form12)
-        if self._debug:
-            unpacked = unpacked1+unpacked2
-            print u' ',zip(self.attrs,unpacked)
-            if len(unpacked) != len(self.attrs):
-                print u' ',unpacked
-
-    def dumpData(self,record,out):
-        """Dumps data from record to outstream."""
-        for target in record.conditions:
-            ##format = 'B3sfI'+target.form12+'4s'
-            out.packSub('CTDA','B3sfI'+target.form12+'4s',
-                target.operFlag, target.unused1, target.compValue,
-                target.ifunc, target.param1, target.param2, target.unused2)
-
-    def mapFids(self,record,function,save=False):
-        """Applies function to fids. If save is true, then fid is set
-        to result of function."""
-        for target in record.conditions:
-            form12 = target.form12
-            if form12[0] == 'I':
-                result = function(target.param1)
-                if save: target.param1 = result
-            if form12[1] == 'I':
-                result = function(target.param2)
-                if save: target.param2 = result
-
 #------------------------------------------------------------------------------
-class MelEffects(MelGroups):
-    """Represents ingredient/potion/enchantment/spell effects."""
-
-    #--Class Data
-    seFlags = bolt.Flags(0x0L,bolt.Flags.getNames('hostile'))
-    class MelEffectsScit(MelStruct):
-        """Subclass to support alternate format."""
-        def __init__(self):
-            MelStruct.__init__(self,'SCIT','II4sB3s',(FID,'script',None),('school',0),
-                ('visual','\x00\x00\x00\x00'),(MelEffects.seFlags,'flags',0x0L),('unused1',null3))
-        def loadData(self,record,ins,type,size,readId):
-            #--Alternate formats
-            if size == 16:
-                attrs,actions = self.attrs,self.actions
-                unpacked = ins.unpack(self.format,size,readId)
-            elif size == 12:
-                attrs,actions = ('script','school','visual'),(0,0,0)
-                unpacked = ins.unpack('II4s',size,readId)
-                record.unused1 = null3
-            else: #--size == 4
-                #--The script fid for MS40TestSpell doesn't point to a valid script.
-                #--But it's not used, so... Not a problem! It's also t
-                record.unused1 = null3
-                attrs,actions = ('script',),(0,)
-                unpacked = ins.unpack('I',size,readId)
-                if unpacked[0] & 0xFF000000L:
-                    unpacked = (0L,) #--Discard bogus MS40TestSpell fid
-            #--Unpack
-            record.__slots__ = self.attrs
-            setter = record.__setattr__
-            for attr,value,action in zip(attrs,unpacked,actions):
-                if callable(action): value = action(value)
-                setter(attr,value)
-            if self._debug: print u' ',unpacked
-
-    #--Instance methods
-    def __init__(self,attr='effects'):
-        """Initialize elements."""
-        MelGroups.__init__(self,attr,
-            MelStruct('EFID','4s',('name','REHE')),
-            MelStruct('EFIT','4s4Ii',('name','REHE'),'magnitude','area','duration','recipient','actorValue'),
-            MelGroup('scriptEffect',
-                MelEffects.MelEffectsScit(),
-                MelString('FULL','full'),
-                ),
-            )
-
 #------------------------------------------------------------------------------
-class MreHasEffects:
-    """Mixin class for magic items."""
-    def getEffects(self):
-        """Returns a summary of effects. Useful for alchemical catalog."""
-        effects = []
-        avEffects = bush.genericAVEffects
-        effectsAppend = effects.append
-        for effect in self.effects:
-            mgef, actorValue = effect.name, effect.actorValue
-            if mgef not in avEffects:
-                actorValue = 0
-            effectsAppend((mgef,actorValue))
-        return effects
-
-    def getSpellSchool(self,mgef_school=bush.mgef_school):
-        """Returns the school based on the highest cost spell effect."""
-        spellSchool = [0,0]
-        for effect in self.effects:
-            school = mgef_school[effect.name]
-            effectValue = bush.mgef_basevalue[effect.name]
-            if effect.magnitude:
-                effectValue *=  effect.magnitude
-            if effect.area:
-                effectValue *=  (effect.area/10)
-            if effect.duration:
-                effectValue *=  effect.duration
-            if spellSchool[0] < effectValue:
-                spellSchool = [effectValue,school]
-        return spellSchool[1]
-
-    def getEffectsSummary(self,mgef_school=None,mgef_name=None):
-        """Return a text description of magic effects."""
-        mgef_school = mgef_school or bush.mgef_school
-        mgef_name = mgef_name or bush.mgef_name
-        with bolt.sio() as buff:
-            avEffects = bush.genericAVEffects
-            aValues = bush.actorValues
-            buffWrite = buff.write
-            if self.effects:
-                school = self.getSpellSchool(mgef_school)
-                buffWrite(bush.actorValues[20+school] + u'\n')
-            for index,effect in enumerate(self.effects):
-                if effect.scriptEffect:
-                    effectName = effect.scriptEffect.full or u'Script Effect'
-                else:
-                    effectName = mgef_name[effect.name]
-                    if effect.name in avEffects:
-                        effectName = re.sub(_(u'(Attribute|Skill)'),aValues[effect.actorValue],effectName)
-                buffWrite(u'o+*'[effect.recipient]+u' '+effectName)
-                if effect.magnitude: buffWrite(u' %sm'%effect.magnitude)
-                if effect.area: buffWrite(u' %sa'%effect.area)
-                if effect.duration > 1: buffWrite(u' %sd'%effect.duration)
-                buffWrite(u'\n')
-                return buff.getvalue()
-
 #------------------------------------------------------------------------------
-class MreLeveledList(MreLeveledListBase):
-    """Leveled item/creature/spell list.."""
-    copyAttrs = ('script','template','chanceNone',)
-
-    #--Special load classes
-    class MelLevListLvld(MelStruct):
-        """Subclass to support alternate format."""
-        def loadData(self,record,ins,type,size,readId):
-            MelStruct.loadData(self,record,ins,type,size,readId)
-            if record.chanceNone > 127:
-                record.flags.calcFromAllLevels = True
-                record.chanceNone &= 127
-
-    class MelLevListLvlo(MelStructs):
-        """Subclass to support alternate format."""
-        def loadData(self,record,ins,type,size,readId):
-            target = self.getDefault()
-            record.__getattribute__(self.attr).append(target)
-            target.__slots__ = self.attrs
-            format,attrs = ((self.format,self.attrs),('iI',('level','listId'),))[size==8]####might be h2sI
-            unpacked = ins.unpack(format,size,readId)
-            setter = target.__setattr__
-            map(setter,attrs,unpacked)
-    #--Element Set
-    melSet = MelSet(
-        MelString('EDID','eid'),
-        MelLevListLvld('LVLD','B','chanceNone'),
-        MelStruct('LVLF','B',(MreLeveledListBase._flags,'flags',0L)),
-        MelFid('SCRI','script'),
-        MelFid('TNAM','template'),
-        MelLevListLvlo('LVLO','h2sIh2s','entries','level',('unused1',null2),(FID,'listId',None),('count',1),('unused2',null2)),
-        MelNull('DATA'),
-        )
-    __slots__ = MreLeveledListBase.__slots__ + melSet.getSlotsUsed()
-
 #------------------------------------------------------------------------------
-class MelOwnership(MelGroup):
-    """Handles XOWN, XRNK, and XGLB for cells and cell children."""
-
-    def __init__(self,attr='ownership'):
-        """Initialize."""
-        MelGroup.__init__(self,attr,
-            MelFid('XOWN','owner'),
-            MelOptStruct('XRNK','i',('rank',None)),
-            MelFid('XGLB','global'),
-        )
-
-    def dumpData(self,record,out):
-        """Dumps data from record to outstream."""
-        if record.ownership and record.ownership.owner:
-            MelGroup.dumpData(self,record,out)
-
 #------------------------------------------------------------------------------
-class MelScrxen(MelFids):
-    """Handles mixed sets of SCRO and SCRV for scripts, quests, etc."""
-
-    def getLoaders(self,loaders):
-        loaders['SCRV'] = self
-        loaders['SCRO'] = self
-
-    def loadData(self,record,ins,type,size,readId):
-        isFid = (type == 'SCRO')
-        if isFid: value = ins.unpackRef(readId)
-        else: value, = ins.unpack('I',4,readId)
-        record.__getattribute__(self.attr).append((isFid,value))
-
-    def dumpData(self,record,out):
-        for isFid,value in record.__getattribute__(self.attr):
-            if isFid: out.packRef('SCRO',value)
-            else: out.packSub('SCRV','I',value)
-
-    def mapFids(self,record,function,save=False):
-        scrxen = record.__getattribute__(self.attr)
-        for index,(isFid,value) in enumerate(scrxen):
-            if isFid:
-                result = function(value)
-                if save: scrxen[index] = (isFid,result)
-
 #------------------------------------------------------------------------------
 # Oblivion Records ------------------------------------------------------------
 #------------------------------------------------------------------------------
