+++ Mopy/bash/game/oblivion/oblivion_const.py
@@ -139,3 +139,179 @@ allBethFiles = set((
     u'Knights.esp',
     u'DLCList.txt',
     ))
+
+# Function Info ---------------------------------------------------------------
+conditionFunctionData = ( #--0: no param; 1: int param; 2: formid param
+    (153, 'CanHaveFlames', 0, 0),
+    (127, 'CanPayCrimeGold', 0, 0),
+    ( 14, 'GetActorValue', 1, 0),
+    ( 61, 'GetAlarmed', 0, 0),
+    (190, 'GetAmountSoldStolen', 0, 0),
+    (  8, 'GetAngle', 1, 0),
+    ( 81, 'GetArmorRating', 0, 0),
+    (274, 'GetArmorRatingUpperBody', 0, 0),
+    ( 63, 'GetAttacked', 0, 0),
+    (264, 'GetBarterGold', 0, 0),
+    (277, 'GetBaseActorValue', 1, 0),
+    (229, 'GetClassDefaultMatch', 0, 0),
+    ( 41, 'GetClothingValue', 0, 0),
+    (122, 'GetCrime', 2, 1),
+    (116, 'GetCrimeGold', 0, 0),
+    (110, 'GetCurrentAIPackage', 0, 0),
+    (143, 'GetCurrentAIProcedure', 0, 0),
+    ( 18, 'GetCurrentTime', 0, 0),
+    (148, 'GetCurrentWeatherPercent', 0, 0),
+    (170, 'GetDayOfWeek', 0, 0),
+    ( 46, 'GetDead', 0, 0),
+    ( 84, 'GetDeadCount', 2, 0),
+    (203, 'GetDestroyed', 0, 0),
+    ( 45, 'GetDetected', 2, 0),
+    (180, 'GetDetectionLevel', 2, 0),
+    ( 35, 'GetDisabled', 0, 0),
+    ( 39, 'GetDisease', 0, 0),
+    ( 76, 'GetDisposition', 2, 0),
+    (  1, 'GetDistance', 2, 0),
+    (215, 'GetDoorDefaultOpen', 0, 0),
+    (182, 'GetEquipped', 2, 0),
+    ( 73, 'GetFactionRank', 2, 0),
+    ( 60, 'GetFactionRankDifference', 2, 2),
+    (128, 'GetFatiguePercentage', 0, 0),
+    (288, 'GetFriendHit', 2, 0),
+    (160, 'GetFurnitureMarkerID', 0, 0),
+    ( 74, 'GetGlobalValue', 2, 0),
+    ( 48, 'GetGold', 0, 0),
+    ( 99, 'GetHeadingAngle', 2, 0),
+    (318, 'GetIdleDoneOnce', 0, 0),
+    (338, 'GetIgnoreFriendlyHits', 0, 0),
+    ( 67, 'GetInCell', 2, 0),
+    (230, 'GetInCellParam', 2, 2),
+    ( 71, 'GetInFaction', 2, 0),
+    ( 32, 'GetInSameCell', 2, 0),
+    (305, 'GetInvestmentGold', 0, 0),
+    (310, 'GetInWorldspace', 2, 0),
+    ( 91, 'GetIsAlerted', 0, 0),
+    ( 68, 'GetIsClass', 2, 0),
+    (228, 'GetIsClassDefault', 2, 0),
+    ( 64, 'GetIsCreature', 0, 0),
+    (161, 'GetIsCurrentPackage', 2, 0),
+    (149, 'GetIsCurrentWeather', 2, 0),
+    (237, 'GetIsGhost', 0, 0),
+    ( 72, 'GetIsID', 2, 0),
+    (254, 'GetIsPlayableRace', 0, 0),
+    (224, 'GetIsPlayerBirthsign', 2, 0),
+    ( 69, 'GetIsRace', 2, 0),
+    (136, 'GetIsReference', 2, 0),
+    ( 70, 'GetIsSex', 1, 0),
+    (246, 'GetIsUsedItem', 2, 0),
+    (247, 'GetIsUsedItemType', 1, 0),
+    ( 47, 'GetItemCount', 2, 0),
+    (107, 'GetKnockedState', 0, 0),
+    ( 80, 'GetLevel', 0, 0),
+    ( 27, 'GetLineOfSight', 2, 0),
+    (  5, 'GetLocked', 0, 0),
+    ( 65, 'GetLockLevel', 0, 0),
+    (320, 'GetNoRumors', 0, 0),
+    (255, 'GetOffersServicesNow', 0, 0),
+    (157, 'GetOpenState', 0, 0),
+    (193, 'GetPCExpelled', 2, 0),
+    (199, 'GetPCFactionAttack', 2, 0),
+    (195, 'GetPCFactionMurder', 2, 0),
+    (197, 'GetPCFactionSteal', 2, 0),
+    (201, 'GetPCFactionSubmitAuthority', 2, 0),
+    (249, 'GetPCFame', 0, 0),
+    (132, 'GetPCInFaction', 2, 0),
+    (251, 'GetPCInfamy', 0, 0),
+    (129, 'GetPCIsClass', 2, 0),
+    (130, 'GetPCIsRace', 2, 0),
+    (131, 'GetPCIsSex', 1, 0),
+    (312, 'GetPCMiscStat', 1, 0),
+    (225, 'GetPersuasionNumber', 0, 0),
+    ( 98, 'GetPlayerControlsDisabled', 0, 0),
+    (365, 'GetPlayerInSEWorld',0,0),
+    (362, 'GetPlayerHasLastRiddenHorse', 0, 0),
+    (  6, 'GetPos', 1, 0),
+    ( 56, 'GetQuestRunning', 2, 0),
+    ( 79, 'GetQuestVariable', 2, 1),
+    ( 77, 'GetRandomPercent', 0, 0),
+    (244, 'GetRestrained', 0, 0),
+    ( 24, 'GetScale', 0, 0),
+    ( 53, 'GetScriptVariable', 2, 1),
+    ( 12, 'GetSecondsPassed', 0, 0),
+    ( 66, 'GetShouldAttack', 2, 0),
+    (159, 'GetSitting', 0, 0),
+    ( 49, 'GetSleeping', 0, 0),
+    ( 58, 'GetStage', 2, 0),
+    ( 59, 'GetStageDone', 2, 1),
+    ( 11, 'GetStartingAngle', 1, 0),
+    ( 10, 'GetStartingPos', 1, 0),
+    ( 50, 'GetTalkedToPC', 0, 0),
+    (172, 'GetTalkedToPCParam', 2, 0),
+    (361, 'GetTimeDead', 0, 0),
+    (315, 'GetTotalPersuasionNumber', 0, 0),
+    (144, 'GetTrespassWarningLevel', 0, 0),
+    (242, 'GetUnconscious', 0, 0),
+    (259, 'GetUsedItemActivate', 0, 0),
+    (258, 'GetUsedItemLevel', 0, 0),
+    ( 40, 'GetVampire', 0, 0),
+    (142, 'GetWalkSpeed', 0, 0),
+    (108, 'GetWeaponAnimType', 0, 0),
+    (109, 'GetWeaponSkillType', 0, 0),
+    (147, 'GetWindSpeed', 0, 0),
+    (154, 'HasFlames', 0, 0),
+    (214, 'HasMagicEffect', 2, 0),
+    (227, 'HasVampireFed', 0, 0),
+    (353, 'IsActor', 0, 0),
+    (314, 'IsActorAVictim', 0, 0),
+    (313, 'IsActorEvil', 0, 0),
+    (306, 'IsActorUsingATorch', 0, 0),
+    (280, 'IsCellOwner', 2, 2),
+    (267, 'IsCloudy', 0, 0),
+    (150, 'IsContinuingPackagePCNear', 0, 0),
+    (163, 'IsCurrentFurnitureObj', 2, 0),
+    (162, 'IsCurrentFurnitureRef', 2, 0),
+    (354, 'IsEssential', 0, 0),
+    (106, 'IsFacingUp', 0, 0),
+    (125, 'IsGuard', 0, 0),
+    (282, 'IsHorseStolen', 0, 0),
+    (112, 'IsIdlePlaying', 0, 0),
+    (289, 'IsInCombat', 0, 0),
+    (332, 'IsInDangerousWater', 0, 0),
+    (300, 'IsInInterior', 0, 0),
+    (146, 'IsInMyOwnedCell', 0, 0),
+    (285, 'IsLeftUp', 0, 0),
+    (278, 'IsOwner', 2, 0),
+    (176, 'IsPCAMurderer', 0, 0),
+    (175, 'IsPCSleeping', 0, 0),
+    (171, 'IsPlayerInJail', 0, 0),
+    (358, 'IsPlayerMovingIntoNewSpace', 0, 0),
+    (339, 'IsPlayersLastRiddenHorse', 0, 0),
+    (266, 'IsPleasant', 0, 0),
+    ( 62, 'IsRaining', 0, 0),
+    (327, 'IsRidingHorse', 0, 0),
+    (287, 'IsRunning', 0, 0),
+    (103, 'IsShieldOut', 0, 0),
+    (286, 'IsSneaking', 0, 0),
+    ( 75, 'IsSnowing', 0, 0),
+    (223, 'IsSpellTarget', 2, 0),
+    (185, 'IsSwimming', 0, 0),
+    (141, 'IsTalking', 0, 0),
+    (265, 'IsTimePassing', 0, 0),
+    (102, 'IsTorchOut', 0, 0),
+    (145, 'IsTrespassing', 0, 0),
+    (329, 'IsTurnArrest', 0, 0),
+    (111, 'IsWaiting', 0, 0),
+    (101, 'IsWeaponOut', 0, 0),
+    (309, 'IsXBox', 0, 0),
+    (104, 'IsYielding', 0, 0),
+    ( 36, 'MenuMode', 1, 0),
+    ( 42, 'SameFaction', 2, 0),
+    (133, 'SameFactionAsPC', 0, 0),
+    ( 43, 'SameRace', 2, 0),
+    (134, 'SameRaceAsPC', 0, 0),
+    ( 44, 'SameSex', 2, 0),
+    (135, 'SameSexAsPC', 0, 0),
+    (323, 'WhichServiceMenu', 0, 0),
+    )
+allConditions = set(entry[0] for entry in conditionFunctionData)
+fid1Conditions = set(entry[0] for entry in conditionFunctionData if entry[2] == 2)
+fid2Conditions = set(entry[0] for entry in conditionFunctionData if entry[3] == 2)
